sort ratting in rad book and wishlist book

+++ src/Component/Utility/getwishList.js
@@ -1,3 +1,5 @@
+import { toast } from "react-toastify";
+
 const getStorwishListBooks = () => {
   const StorWishbook = localStorage.getItem("Wish-Book");
 
@@ -8,8 +10,13 @@ const getStorwishListBooks = () => {
 };
 const SaveWishListBook = (book) => {
   const StorWishBooks = getStorwishListBooks();
-
-  StorWishBooks.push(book);
-  localStorage.setItem("Wish-Book", JSON.stringify(StorWishBooks));
+  const exist = StorWishBooks.find((itms) => itms.id === book.id);
+  if (exist) {
+    return toast.error("Alrady added in local Storag");
+  } else {
+    StorWishBooks.push(book);
+    localStorage.setItem("Wish-Book", JSON.stringify(StorWishBooks));
+    toast.success("Added wishlist successfilly ");
+  }
 };
 export { getStorwishListBooks, SaveWishListBook };
